# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Vellaveto — MCP Tool Firewall API
  description: |
    Runtime security engine for AI agent tool calls. Intercepts MCP
    (Model Context Protocol) and function-calling requests, enforces
    security policies on paths/domains/actions, and maintains a
    tamper-evident audit trail.

    ## Authentication

    All endpoints (except `/health`, `/setup/*`, `/iam/*`, and billing
    webhooks) require one of:
    - **API Key** — `Authorization: Bearer <key>` header
    - **OAuth 2.0** — Bearer token from OIDC provider

    ## Multi-Tenancy

    Pass `X-Tenant-ID: <tenant>` header to scope requests to a tenant.
    Default tenant is used when the header is absent.

    ## Rate Limiting

    All endpoints are rate-limited. Categories:
    - **evaluate** — higher throughput (default 1000 req/s)
    - **admin** — moderate (default 100 req/s)
    - **readonly** — moderate (default 200 req/s)

    HTTP 429 is returned when limits are exceeded.
  version: 5.0.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Vellaveto
    url: https://github.com/paolovella/vellaveto
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://vellaveto.example.com
    description: Production
...
security:
  - apiKey: []
  - oauth2: []
paths:
...
  /api/evaluate:
    post:
      operationId: evaluate
      summary: Evaluate an action against policies
      description: |
        Core endpoint. Evaluates a tool call action against loaded policies
        and returns a verdict (Allow, Deny, or RequireApproval). The action's
        parameters are redacted in the response for security.
      tags: [Evaluation]
      parameters:
        - name: trace
          in: query
          schema:
            type: boolean
            default: false
          description: Enable detailed evaluation trace in response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

  # ═══════════════════════════════════════════════════════════════════════════
  # POLICIES
  # ═══════════════════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: API key passed as Bearer token
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /iam/login
          tokenUrl: /iam/callback
          scopes:
            read: Read access
            write: Write access
            admin: Admin access
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
...
    Action:
      type: object
      required: [tool, function]
      properties:
        tool:
          type: string
          description: Tool name (e.g., "file_read", "api_gateway")
          maxLength: 256
        function:
          type: string
          description: Function name (e.g., "read", "query")
          maxLength: 256
        parameters:
          type: object
          additionalProperties: true
          description: Tool call parameters (key-value pairs)
        target_paths:
          type: array
          items:
            type: string
          description: File/resource paths accessed by this action
        target_domains:
          type: array
          items:
            type: string
          description: Network domains accessed by this action
        resolved_ips:
          type: array
          items:
            type: string
          description: Resolved IP addresses
    Verdict:
      oneOf:
        - type: string
          enum: [Allow]
        - type: object
          properties:
            Deny:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
        - type: object
          properties:
            RequireApproval:
              type: object
              properties:
                approval_id:
                  type: string
...
    EvaluationContext:
      type: object
      properties:
        agent_id:
          type: string
        session_id:
          type: string
        call_chain:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties:
            type: string

    # ─── Request/Response Types ─────────────────────────────────────────
    EvaluateRequest:
      allOf:
        - $ref: "#/components/schemas/Action"
        - type: object
          properties:
            context:
              $ref: "#/components/schemas/EvaluationContext"
    EvaluateResponse:
      type: object
      required: [verdict, action]
      properties:
        verdict:
          $ref: "#/components/schemas/Verdict"
        action:
          $ref: "#/components/schemas/Action"
          description: Action with parameters redacted
        approval_id:
          type: string
          description: Present when verdict is RequireApproval
        trace:
          type: object
          description: Detailed evaluation trace (when trace=true)
...